refactored code so that the directory is more readable
